+++ releasenotes.md
@@ -21,6 +21,7 @@
 - Indexing an Optional slice would crash in codegen #1636.
 - SimpleHeapAllocator bug when splitting blocks allowed memory overrun.
 - Not possible to alias or take reference for extension methods on non-user defined types. #1637
+- Prevent methods from using names of properties or fields. #1638
 
 ### Stdlib changes
 - Add `io::MultiReader`, `io::MultiWriter`, and `io::TeeReader` structs.